requirements.txt
flask
requests
flask-socketio
flask-cors
Pillow
flask-jwt-extended
pytz
pyzbar
eventlet
pymysql
flask_sqlalchemy
python-dotenv
flask-login
bcrypt
qrcode
line-bot-sdk
pypromptpay
segno
